refactor(shared): move structural workflow validation into workflow-graph (stage 12-4)
