+++ apps/worker/package.json
@@ -15,6 +15,7 @@
     "validate:pre-sandbox": "tsx scripts/validate-pre-sandbox-config.ts",
     "validate:local-skills": "tsx scripts/validate-local-skills.ts",
     "capture:agent-protocol-fixtures": "tsx scripts/capture-agent-protocol-fixtures.ts",
+    "capture:definition-golden": "tsx scripts/capture-definition-issue-golden.ts",
     "cleanup:startup-orphans": "tsx scripts/cleanup-startup-orphans.ts",
     "cleanup:orphaned-running-runs": "tsx scripts/cleanup-orphaned-running-runs.ts",
     "mcp:smoke": "tsx scripts/mcp-smoke.ts",
