+++ apps/worker/src/workflow-definition/scenarios/schedule-open-pr.scenario.test.ts
@@ -7,9 +7,7 @@ import {
   testBlockData,
   testDeploymentIssues,
 } from "../../test-support/block-contracts.js";
-import {
-  workflowDefinitionV2Schema,
-} from "../schema.js";
+import { workflowDefinitionV2Schema } from "@shared/workflow-graph";
 import { executorRunsOf, expectStartsAfterFinishOf } from "./assertions.js";
 import { createScenario, type Scenario } from "./harness.js";
