+++ packages/workflow-graph/index.ts
@@ -1,2 +1,5 @@
+export * from "./graph-issues";
+export * from "./limits";
+export * from "./schema";
 export * from "./v2-bindings";
 export * from "./v2-branch";
