+++ scripts/ci/verify-changed.ts
@@ -25,6 +25,9 @@ export const WORKFLOW_TESTS = [
 export const WORKFLOW_GRAPH_TESTS = [
   "src/engine/workflow-import-boundary.test.ts",
   "src/routes/import-graph-guard.test.ts",
+  "src/mcp/tool-catalog.test.ts",
+  "src/workflow-definition/__golden__/definition-deployment-issues.test.ts",
+  "src/workflow-definition/schema-v2.test.ts",
   "src/workflow-definition/v2-bindings.test.ts",
   "src/workflow-definition/v2-branch.test.ts",
 ] as const;
